requirements.txt
aniso8601==9.0.1
attrs==22.1.0
blinker==1.5
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
flask-mongoengine==1.0.0
Flask-Principal==0.4.0
flask-restx==0.5.1
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.13.0
MarkupSafe==2.1.1
mongoengine==0.24.2
pymongo==4.2.0
pyrsistent==0.18.1
pytz==2022.2.1
six==1.16.0
Werkzeug==2.1.2
WTForms==3.0.1

python-dateutil~=2.8.2